refactor: split url package types into separate files
