+++ internal/pkg/url/url.go
@@ -19,135 +19,63 @@
 package url
 
 import (
-	"encoding/base64"
 	"fmt"
-	"net/http"
-	"os"
-	"path/filepath"
 	"regexp"
+	"slices"
 	"strings"
-
-	"github.com/jdx/go-netrc"
 )
 
-type BasicAuth struct {
-	Username, Password string
-}
-
-func (auth *BasicAuth) Encode() string {
-	if auth == nil {
-		return ""
+type (
+	ParsedURL struct {
+		Scheme   string
+		Username string
+		Password string
+		Hostname string
+		Port     string
+		GitRef   string
+		Path     string
+		Query    string
+		Fragment string
+		Tag      string
+		Digest   string
 	}
 
-	data := []byte(strings.Join([]string{auth.Username, auth.Password}, ":"))
-
-	return base64.URLEncoding.EncodeToString(data)
-}
-
-func (*BasicAuth) Name() string {
-	return "http-basic-auth"
-}
-
-func (auth *BasicAuth) SetAuth(request *http.Request) {
-	if auth == nil {
-		return
+	Parser interface {
+		Parse(fetchURL string) *ParsedURL
+		RegExp() *regexp.Regexp
 	}
+)
 
-	request.SetBasicAuth(auth.Username, auth.Password)
-}
-
-func (auth *BasicAuth) String() string {
-	masked := "*******"
-	if auth.Password == "" {
-		masked = "<empty>"
-	}
-
-	return fmt.Sprintf("Authorization: Basic %s:%s", auth.Username, masked)
-}
-
-func (auth *BasicAuth) UseNetRC(hostname string) error {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return fmt.Errorf("failed to get user home directory: %w", err)
-	}
-
-	authFile, err := netrc.Parse(filepath.Join(home, ".netrc"))
-	if err != nil {
-		return fmt.Errorf("failed to parse .netrc file: %w", err)
-	}
-
-	// Use credentials in .netrc if entry for the hostname is found
-	if machine := authFile.Machine(hostname); machine != nil {
-		auth.Username = machine.Get("login")
-		auth.Password = machine.Get("password")
-	}
-
-	return nil
-}
-
-type ParsedURL struct {
-	Scheme   string
-	Username string
-	Password string
-	Hostname string
-	Port     string
-	GitRef   string
-	Path     string
-	Query    string
-	Fragment string
-	Tag      string
-	Digest   string
-}
-
-func (url *ParsedURL) String() string { //nolint:cyclop
-	var (
-		urlBytes []byte
-		pathSep  string
-	)
+func (url *ParsedURL) String() string {
+	var urlString, pathSep string
 
 	switch url.Scheme {
 	case "http", "https", "oci":
-		urlBytes = append(urlBytes, fmt.Sprintf("%s://", url.Scheme)...)
+		urlString += fmt.Sprintf("%s://", url.Scheme)
 		pathSep = "/"
 	case "git", "ssh":
-		urlBytes = append(urlBytes, fmt.Sprintf("%s@", url.Scheme)...)
+		urlString += fmt.Sprintf("%s@", url.Scheme)
 		pathSep = ":"
 	}
 
 	if url.Username != "" && url.Password != "" {
-		urlBytes = append(urlBytes, fmt.Sprintf("%s:%s@", url.Username, url.Password)...)
+		urlString += fmt.Sprintf("%s:%s@", url.Username, url.Password)
 	}
 
-	urlBytes = append(urlBytes, url.Hostname...)
+	urlString += url.Hostname
 
-	if url.Path != "" {
-		urlBytes = append(urlBytes, fmt.Sprintf("%s%s", pathSep, url.Path)...)
+	removeEmpty := func(s string) []string {
+		return slices.DeleteFunc(
+			[]string{urlString, s}, func(s string) bool { return s == "" },
+		)
 	}
 
-	if url.GitRef != "" {
-		urlBytes = append(urlBytes, fmt.Sprintf("@%s", url.GitRef)...)
-	}
+	urlString = strings.Join(removeEmpty(url.Path), pathSep)
+	urlString = strings.Join(removeEmpty(url.GitRef), "@")
+	urlString = strings.Join(removeEmpty(url.Query), "?")
+	urlString = strings.Join(removeEmpty(url.Fragment), "#")
+	urlString = strings.Join(removeEmpty(url.Tag), ":")
+	urlString = strings.Join(removeEmpty(url.Digest), "@")
 
-	if url.Query != "" {
-		urlBytes = append(urlBytes, fmt.Sprintf("?%s", url.Query)...)
-	}
-
-	if url.Fragment != "" {
-		urlBytes = append(urlBytes, fmt.Sprintf("#%s", url.Fragment)...)
-	}
-
-	if url.Tag != "" {
-		urlBytes = append(urlBytes, fmt.Sprintf(":%s", url.Tag)...)
-	}
-
-	if url.Digest != "" {
-		urlBytes = append(urlBytes, fmt.Sprintf("@%s", url.Digest)...)
-	}
-
-	return string(urlBytes)
-}
-
-type Parser interface {
-	Parse(fetchURL string) *ParsedURL
-	RegExp() *regexp.Regexp
+	return urlString
 }
